WebStorage/tests/approved/storage_local_setitem.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_local_setItem</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('localStorage' in window) && window.localStorage !== null){
			localStorage.clear();

			try {
				localStorage.setItem("name", "user1");
			} catch (e) {
				test(function()
				{
					assert_unreached("setItem method is failed.");
				}, "setItem method test")
			}

			test(function() {
				assert_not_equals(localStorage.length, 0, "localStorage.setItem")
			}, "setItem method test")

		} else {
			test(function()
			{
				assert_unreached("localStorage is not supported.");
			}, "localStorage support test")
		}	
	</script>

 </body>
</html>
